fix: SwiftLint file_length 違反を解消する
TwitchIRCClientRateLimitTests を独立ファイルに移動して
TwitchIRCClientTests.swift の行数を 702 → 635 行に削減する（上限: 700行）
